change other attributor and ensemble
