+++ example/cifar2_resnet9/influence_function_noisy_label.py
@@ -6,8 +6,6 @@ from functools import partial
 
 import torch
 from torch import nn
-from torchvision import transforms
-
 from dattri.algorithm.influence_function import\
     IFAttributorExplicit,\
     IFAttributorCG,\
