+++ example/cifar2_resnet9/tracin_noisy_label.py
@@ -2,36 +2,19 @@
 
 import torch
 from torch import nn
-from torch.utils.data import Sampler
-from torchvision import transforms
 
 from dattri.algorithm.tracin import TracInAttributor
 from dattri.benchmark.datasets.cifar2 import create_cifar2_dataset, train_cifar2_resnet9
-from dattri.benchmark.utils import flip_label
-from dattri.func.utils import flatten_func
+from dattri.benchmark.utils import flip_label, SubsetSampler
+from dattri.task import AttributionTask
 
 
 def get_cifar_indices_and_adjust_labels(dataset, subset_indice):
-
     dataset.targets, flip_index = flip_label(torch.tensor(dataset.targets)[subset_indice], p=0.1)
     return flip_index
 
 
-class SubsetSampler(Sampler):
-    def __init__(self, indices):
-        self.indices = indices
-
-    def __iter__(self):
-        return iter(self.indices)
-
-    def __len__(self):
-        return len(self.indices)
-
-
 if __name__ == "__main__":
-    transform = transforms.Compose([transforms.ToTensor(),
-                                    transforms.Normalize((0.5, 0.5, 0.5),
-                                                         (0.5, 0.5, 0.5))])
     train_dataset, _ = create_cifar2_dataset("./data")
 
     train_loader = torch.utils.data.DataLoader(
@@ -54,7 +37,6 @@ if __name__ == "__main__":
     model_1.cuda()
     model_1.eval()
 
-    @flatten_func(model_1)
     def f(params, image_label_pair):
         image, label = image_label_pair
         image_t = image.unsqueeze(0)
@@ -63,11 +45,15 @@ if __name__ == "__main__":
         yhat = torch.func.functional_call(model_1, params, image_t)
         return loss(yhat, label_t)
 
-    # simulate checkpoints
-    model_params_1 = {k: p for k, p in model_1.named_parameters() if p.requires_grad}
-    attributor = TracInAttributor(
+    task = AttributionTask(
         target_func=f,
-        params_list=[model_params_1],
+        model=model_1,
+        checkpoints=model_1.state_dict(),
+    )
+
+    # simulate checkpoints
+    attributor = TracInAttributor(
+        task=task,
         weight_list=torch.tensor([0.01]),
         normalized_grad=True,
         device=torch.device("cuda"),
