+++ test/dattri/algorithm/test_trak.py
@@ -9,7 +9,7 @@ from torch.utils.data import TensorDataset
 
 from dattri.algorithm.trak import TRAKAttributor
 from dattri.benchmark.datasets.mnist import train_mnist_lr
-from dattri.func.utils import flatten_func
+from dattri.task import AttributionTask
 
 
 class TestTRAK:
@@ -24,7 +24,6 @@ class TestTRAK:
 
         model = train_mnist_lr(train_loader)
 
-        @flatten_func(model)
         def f(params, image_label_pair):
             image, label = image_label_pair
             image_t = image.unsqueeze(0)
@@ -33,7 +32,6 @@ class TestTRAK:
             yhat = torch.func.functional_call(model, params, image_t)
             return loss(yhat, label_t)
 
-        @flatten_func(model)
         def m(params, image_label_pair):
             image, label = image_label_pair
             image_t = image.unsqueeze(0)
@@ -56,13 +54,12 @@ class TestTRAK:
             "device": "cpu",
             "use_half_precision": False,
         }
+        task = AttributionTask(target_func=f, model=model, checkpoints=checkpoint_list)
 
         # trak w/o cache
         attributor = TRAKAttributor(
-            f,
-            m,
-            model=model,
-            checkpoint_list=["ckpts/model_1.pt"],
+            task=task,
+            correct_probability_func=m,
             device=torch.device("cpu"),
             projector_kwargs=projector_kwargs,
         )
